pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "subtitles2text"
version = "0.0.1"
dependencies = [
    "pm-vtt2txt @ git+https://github.com/lrq3000/pm-vtt2txt",
]

[project.scripts]
subtitles2text = "subtitles2text.main:run_gui"